refactoring source/new-item-in-feed, action/merge-rss-feeds, rss.app moving commonly used methods to rss.app moving from .js to .mjs
