--- components/rss/rss.app.js
@@ -1,4 +1,0 @@
-module.exports = {
-  type: 'app',
-  app: 'rss',
-}
